// src/helpers/interfaces/TicketModel.ts
export interface TicketModel {
    ticketId: string;
    isTicketSolved: boolean;
    target_team: string;
    title: string;
    Priority: string;
    submitted: string;
    updatedAt: string;
    solvedBy: string;
    description: string;
    issueRaisedBy: string;
    ticketType: string;
    otherInfo: string;
    createdAt: string;
    ticketActionCode: string;
    comments: string;
    solved_by: string;
  }
  